Move example bank app into a test fixture
