+++ example/bank/bin/transfer
@@ -2,7 +2,7 @@
 
 const uuidv4 = require('uuid')
 
-const {TRANSFER} = require('../src/command.js')
+const {TRANSFER} = require('../../../test/fixture/app/bank/command.js')
 const {accountIdByName, die, executeCommands, runAsync, withClient} = require('./util.js')
 
 runAsync(withClient(async pgClient => {
