--- example/bank/src/aggregate/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  account: require('./account.js'),
-  transaction: require('./transaction.js'),
-}
