--- example/bank/src/initialize.js
@@ -1,5 +1,0 @@
-const {initializeSchema} = require('./schema.js')
-
-module.exports = async function initialize (pgClient) {
-  await initializeSchema(pgClient)
-}
