--- example/bank/src/process/index.js
@@ -1,5 +1,0 @@
-module.exports = {
-  deposit: require('./deposit.js'),
-  transfer: require('./transfer.js'),
-  withdrawal: require('./withdrawal.js'),
-}
