--- example/bank/src/projection/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  relational: require('./relational.js'),
-}
