+++ test/fixture/app/bank/aggregate/transaction.js
