+++ test/fixture/app/bank/command.js
