+++ test/fixture/app/bank/initialize.js
@@ -1,0 +1,7 @@
+const {initializeSchema} = require('./schema.js')
+
+module.exports = function createInitialize (schemaName) {
+  return async function initialize (pgClient) {
+    await initializeSchema(pgClient, schemaName)
+  }
+}
